requirements.txt
numpy
requests
pillow
BytesIO